Only use a sticky header if we're using scrolling animations
Animating the header automatically is super annoying once you've seen it
once and want to scroll past it. We could either always have a
functional sticky header on larger screens, but we already do away with
the sticky header on small screens in favour of the much more dramatic
splash image, so we take the same approach is we can't do the more fun
scrolling animation.